Move classic editor processing code to a Controller
