+++ src/Core/HooksAbstract.php
@@ -22,4 +22,5 @@ abstract class HooksAbstract
     const ACTION_BULK_EDIT_CUSTOM_BOX = 'bulk_edit_custom_box';
     const ACTION_QUICK_EDIT_CUSTOM_BOX = 'quick_edit_custom_box';
     const ACTION_SAVE_POST = 'save_post';
+    const ACTION_ADD_META_BOX = 'add_meta_boxes';
 }
